Meter & Building Bounce
Fixed Hud with the rhythm filling and the building/scenery bouncing. I
haven't testing it, but I was getting some major frame rate issues on my
end towards the end, might have too many things moving at once, or it
might be too complex. Might have to write a script to go with it that
actiavtes/deactivates the bounce based on player position. Let me know
if it is lagging up for you guys, if so i'll look into it.